+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 * Fix: mounting multiple documentations - [@Drakula2k](https://github.com/Drakula2k).
 * Fix: resource groupings for prefixed APIs - [@aew](https://github.com/aew).
 * Fix: hide_documentation_path on prefixed APIs - [@spier](https://github.com/spier).
+* [#87](https://github.com/tim-vandecasteele/grape-swagger/pull/87) - Fix: mapping of `default` to `defaultValue` - [@m-o-e](https://github.com/m-o-e).
 * Rewritten .gemspec and removed Jeweler - [@dblock](https://github.com/dblock).
 * Added `GrapeEntity::VERSION` - [@dblock](https://github.com/dblock).
 * Added Rubocop, Ruby-style linter - [@dblock](https://github.com/dblock).
+
 * Your Contribution Here
 
 ### 0.6.0 (June 19, 2013)